+++ ketos/tests/data_handling/test_selection_table.py
@@ -334,6 +334,27 @@ def test_select_step(annot_table_std):
     M = len(df_new)
     assert M == (N - K) * (2 * int((3.3/2+0.5-0.4)/0.5) + 1) + K * (2 * int((3.3/2-0.5)/0.5) + 1)
 
+def test_select_warning_annotation_error(annot_table_std):
+    """ Test that a warning is given when an annotation contains a start time greater than end time
+    and that select ignores the annotation
+    """
+    np.random.seed(2)
+    df = annot_table_std
+    df["end"][1] = 0.23
+    df = st.standardize(df, start_labels_at_1=True)
+
+    with pytest.warns(UserWarning):
+        res = st.select(annotations=df, length=1.0, center=False)
+
+    d ='''filename sel_id label start end                            
+f0.wav   0           3  1.002788  2.002788
+f0.wav   1           2  3.059630  4.059630
+f1.wav   1           2  5.264224  6.264224
+f2.wav   0           5  3.001242  4.001242
+f2.wav   1           1  5.966846  6.966846'''
+    ans = pd.read_csv(StringIO(d), delim_whitespace=True, index_col=[0,1])
+    pd.testing.assert_frame_equal(ans, res[ans.columns.values])
+
 
 def test_time_shift():
     row = pd.Series({'label':3.00,'start':0.00,'end':3.30,'annot_id':0.00,'length':3.30,'start_new':-0.35})
@@ -347,6 +368,23 @@ def test_time_shift():
     ans = pd.read_csv(StringIO(d), delim_whitespace=True, index_col=[0])
     pd.testing.assert_frame_equal(ans, res[ans.columns.values])
 
+def test_time_shift_negative_length_raises_exception():
+    """ Test that the assertion error for a length smaller than 0 is thrown
+    """
+    annot = {'filename':'file1.wav', 'label':1, 'start':12.0, 'end':14.0}
+
+    with pytest.raises(AssertionError):
+        st.time_shift(annot, time_ref=13.0, length=-1.0, step=0.2, min_overlap=0.5)
+
+def test_time_shift_positive_length_doesnt_raises_exception():
+    """ Test that the assertion error is not thrown for good values
+    """
+    annot = {'filename':'file1.wav', 'label':1, 'start':12.0, 'end':14.0}
+
+    try:
+        st.time_shift(annot, time_ref=13.0, length=1.0, step=0.2, min_overlap=0.5)
+    except AssertionError as exc:
+        assert False, f"'time_shift' raised an exception {exc}"
 
 def test_select_with_varying_overlap(annot_table_std):
     """ Test that the number of selections increases as the